Refaktorer GUI i egne moduler
